Animations
Added the spine animations for the Cleric and the Succubus, all of them, Yes. I did the job, Now stop crying and let me be,  I want to rest but noooo, I must MAKE THE NEXT ROL SESSION FOR MY OTHER BOIS CUZ IS FOR THIS SUNDAY MORNING AND I ONLY HAVE 12 HOURS TO BE READY AND I ABSOLUTELY DONT HAVE ANXIAETY  AT ALL!! oh, and Hello Fire, wassup?