Particles half implemented
Added a particle effect to the healing animation and added missing frames to the Mace animation. Now Noob you should fix the animations because they are laggy as frick and that's vewy bawd.

TODO: Make a custom animation for the blinding light spell with particles and maybe a particle effect for the Kiss animation